# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cleaning company app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /houses:
    get:
      operationId: listHouses
      tags: [houses]
      summary: List all houses
      responses:
        "200":
          description: List of houses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/House"
    post:
      operationId: createHouse
      tags: [houses]
      summary: Create a new house
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHouseBody"
      responses:
        "201":
          description: Created house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
  /houses/{id}:
    get:
      operationId: getHouse
      tags: [houses]
      summary: Get a single house
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: House
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
    put:
      operationId: updateHouse
      tags: [houses]
      summary: Update a house
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseBody"
      responses:
        "200":
          description: Updated house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
    delete:
      operationId: deleteHouse
      tags: [houses]
      summary: Delete a house
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /houses/{id}/notes:
    patch:
      operationId: updateHouseNotes
      tags: [houses]
      summary: Update notes for a house (editable by anyone)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseNotesBody"
      responses:
        "200":
          description: Updated house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
...
components:
  schemas:
...
    House:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mapLink:
          type: string
          nullable: true
        ownerName:
          type: string
        ownerPhone:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        singleBeds:
          type: integer
          nullable: true
        doubleBeds:
          type: integer
          nullable: true
        babyBeds:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        jacuzzis:
          type: integer
          nullable: true
        saunas:
          type: integer
          nullable: true
        entryCode:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
      required:
        - id
        - name
        - ownerName
        - status
        - createdAt
    CreateHouseBody:
      type: object
      properties:
        name:
          type: string
        mapLink:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
        ownerPhone:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        singleBeds:
          type: integer
          nullable: true
        doubleBeds:
          type: integer
          nullable: true
        babyBeds:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        jacuzzis:
          type: integer
          nullable: true
        saunas:
          type: integer
          nullable: true
        entryCode:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
      required:
        - name
        - status
    UpdateHouseBody:
      type: object
      properties:
        name:
          type: string
        mapLink:
          type: string
          nullable: true
        ownerName:
          type: string
        ownerPhone:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        singleBeds:
          type: integer
          nullable: true
        doubleBeds:
          type: integer
          nullable: true
        babyBeds:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        jacuzzis:
          type: integer
          nullable: true
        saunas:
          type: integer
          nullable: true
        entryCode:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
...
    UpdateHouseNotesBody:
      type: object
      properties:
        notes:
          type: string
          nullable: true
      required:
        - notes
...