# lib/db/src/schema/houses.ts
import { pgTable, serial, text, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { companiesTable } from "./companies";

export const housesTable = pgTable("houses", {
  id: serial("id").primaryKey(),
  companyId: integer("company_id").references(() => companiesTable.id),
  name: text("name").notNull(),
  mapLink: text("map_link"),
  ownerName: text("owner_name").notNull(),
  ownerPhone: text("owner_phone"),
  ownerEmail: text("owner_email"),
  notes: text("notes"),
  singleBeds: integer("single_beds"),
  doubleBeds: integer("double_beds"),
  babyBeds: integer("baby_beds"),
  bathrooms: integer("bathrooms"),
  jacuzzis: integer("jacuzzis"),
  saunas: integer("saunas"),
  entryCode: text("entry_code"),
  status: text("status").notNull().default("active"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertHouseSchema = createInsertSchema(housesTable).omit({ id: true, createdAt: true });
export type InsertHouse = z.infer<typeof insertHouseSchema>;
export type House = typeof housesTable.$inferSelect;